+++ test/secrets_test.go
@@ -15,10 +15,10 @@ import (
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
 	clicmd "github.com/supabase/cli/cmd"
-	"github.com/supabase/cli/integration/mocks/supabase"
+	"github.com/supabase/cli/test/mocks/supabase"
 )
 
-type LinkTestSuite struct {
+type SecretsTestSuite struct {
 	suite.Suite
 	tempDir string
 	cmd     *cobra.Command
@@ -30,34 +30,44 @@ type LinkTestSuite struct {
 }
 
 // test functions
-func (suite *LinkTestSuite) TestLink() {
+func (suite *SecretsTestSuite) TestLink() {
 	// run command
-	link, _, err := suite.cmd.Find([]string{"link"})
+	list, _, err := suite.cmd.Find([]string{"secrets", "list"})
 	require.NoError(suite.T(), err)
+
 	key := "sbp_" + gonanoid.MustGenerate(supabase.KeyAlphabet, supabase.KeyLength)
 	os.Setenv("SUPABASE_ACCESS_TOKEN", key)
-	id := gonanoid.MustGenerate(supabase.IDAlphabet, supabase.IDLength)
-	require.NoError(suite.T(), link.Flags().Set("project-ref", id))
-	require.NoError(suite.T(), link.RunE(link, []string{}))
+	ref := gonanoid.MustGenerate(supabase.IDAlphabet, supabase.IDLength)
+	require.NoError(suite.T(), ioutil.WriteFile("supabase/.temp/project-ref", []byte(ref), os.FileMode(0755)))
+
+	// set stdout to write into file so we can capture cmd output
+	tmpfile, err := ioutil.TempFile(suite.tempDir, "output")
+	require.NoError(suite.T(), err)
+	defer os.Remove(tmpfile.Name()) // clean up
+	oldStdout := os.Stdout
+	defer func() { os.Stdout = oldStdout }()
+	os.Stdout = tmpfile
+
+	require.NoError(suite.T(), list.RunE(list, []string{}))
 
 	// check request details
 	suite.mtx.RLock()
 	defer suite.mtx.RUnlock()
-	require.Contains(suite.T(), suite.ids, id)
+	require.Contains(suite.T(), suite.ids, ref)
 	require.Contains(suite.T(), suite.headers, http.Header{
 		"Authorization":   []string{fmt.Sprintf("Bearer %s", key)},
 		"Accept-Encoding": []string{"gzip"},
 		"User-Agent":      []string{"Go-http-client/1.1"},
 	})
-	_, err = os.Stat("supabase/.temp/project-ref")
+
+	contents, err := ioutil.ReadFile(tmpfile.Name())
 	require.NoError(suite.T(), err)
-	ref, err := ioutil.ReadFile("supabase/.temp/project-ref")
-	require.NoError(suite.T(), err)
-	require.Equal(suite.T(), id, string(ref))
+	require.Contains(suite.T(), string(contents), "some-key")
+	require.Contains(suite.T(), string(contents), "another")
 }
 
 // hooks
-func (suite *LinkTestSuite) SetupTest() {
+func (suite *SecretsTestSuite) SetupTest() {
 	// init cli
 	suite.cmd = clicmd.GetRootCmd()
 	suite.tempDir = NewTempDir(Logger, TempDir)
@@ -67,33 +77,45 @@ func (suite *LinkTestSuite) SetupTest() {
 	require.NoError(suite.T(), err)
 	require.NoError(suite.T(), init.RunE(init, []string{}))
 
+	// add `link` dir
+	require.NoError(suite.T(), os.MkdirAll("supabase/.temp", os.FileMode(0755)))
+
 	// implement mocks
-	SupaMock.FunctionsHandler = func(c *gin.Context) {
+	SupaMock.SecretsHandler = func(c *gin.Context) {
 		suite.addHeaders(c.Request.Header)
 		suite.addID(c.Params.ByName("id"))
 
-		c.JSON(http.StatusOK, gin.H{})
+		c.JSON(http.StatusOK, []gin.H{
+			{
+				"Name":  "some-key",
+				"Value": gonanoid.Must(),
+			},
+			{
+				"Name":  "another",
+				"Value": gonanoid.Must(),
+			},
+		})
 	}
 }
 
-func (suite *LinkTestSuite) TeardownTest() {
+func (suite *SecretsTestSuite) TeardownTest() {
 	require.NoError(suite.T(), os.Chdir(TempDir))
 }
 
 // In order for 'go test' to run this suite, we need to create
 // a normal test function and pass our suite to suite.Run
-func TestLinkTestSuite(t *testing.T) {
-	suite.Run(t, new(LinkTestSuite))
+func TestSecretsTestSuite(t *testing.T) {
+	suite.Run(t, new(SecretsTestSuite))
 }
 
 // helper functions
-func (suite *LinkTestSuite) addID(id string) {
+func (suite *SecretsTestSuite) addID(id string) {
 	suite.mtx.Lock()
 	defer suite.mtx.Unlock()
 	suite.ids = append(suite.ids, id)
 }
 
-func (suite *LinkTestSuite) addHeaders(headers http.Header) {
+func (suite *SecretsTestSuite) addHeaders(headers http.Header) {
 	suite.mtx.Lock()
 	defer suite.mtx.Unlock()
 	suite.headers = append(suite.headers, headers)
